scripts/requirements.txt
appnope==0.1.4
asttokens==2.4.1
blinker==1.8.2
click==8.1.7
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
Faker==30.8.1
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
wcwidth==0.2.13
Werkzeug==3.1.0
